pyproject.toml
[project]
name = "orchael-sdk"
version = "0.1.0"
description = "A framework for building chat processors and chat applications"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Orchael Team"}
]
license = {text = "MIT"}
keywords = ["chat", "processor", "sdk", "framework"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "typing-extensions>=4.0.0",
    "click>=8.0.0",
    "PyYAML>=6.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0"
]

[project.scripts]
orchael-sdk-cli = "orchael_sdk.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchael_sdk"]

[tool.hatch.build.targets.sdist]
include = [
    "/orchael_sdk",
    "/example",
    "/example2", 
    "/examples",
    "/README.md",
    "/pyproject.toml"
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py310"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]
